Experimental support for session-backed queries. See #1040.
Information about the most recently executed query gets stored in the users session. The next/previous links on ticket pages then move backward/forward in the query result set instead of globally. Also, there's a link to go back to the query screen. The query results display will highlight tickets that have been added or changed since the first execution. Tickets that have been changed so that they no longer match the constraints of the query are still shown but grayed out. I'll need to check whether some of this functionality can be ported to reports.