''xref-branch'': basic objectification done.
There's now an {{{object_factory}}} modeled after the {{{module_factory}}}.
The !WikiPage and Ticket classes are now also inheriting !TracObj.

Small simplification of the presentation of cross-references.

From this point, I'll create the ''trac-obj-branch'' to go on
with the more advanced stuff about objectification of Trac,
and the ''xref-branch'' will start its (hopefully brief!) way
towards acceptance in ''trunk''.