Enhanced Revision Log.
The normal ''node'' history now highlights the add/modify/copy events (#1428)
and there's a new alternate revision log that focus on the ''path'' history.
The two are well integrated and one can for example easily go to the ''node''
history of any of the copy change listed in the path history, in case there
has been multiple add/delete or copy/delete on that path.

This implements #1441.