TicketTypes: diff showing what would be a renaming of {{{ticket_type}}} to {{{type}}}
This renaming would introduce amgiguities...
Some of them are easily worked around (in Timeline.py/timeline.cs),
while the ambiguity introduced in the SQL generated by the Query module is a differenbeast:
{{{type}}} is also a column name for the {{{enum}}} and {{{ticket_custom}}} tables,
and the SQL engine (at least SQLite) complains when it sees the unqualified {{{type}}} column name.

The possible solutions could be:
 * to check everywhere for the special case {{{type <-> ticket.type}}} 
   (as the code implies equivalence of the filter name and the ticket column name), 
   which is not nice,
 * add the {{{ticket.}}} prefix systematically, as this patch tries to do,
   but there are still corner cases (esp. related to custom fields) which
   do not work...
 * use {{{kind}}} instead of {{{type}}} for the {{{enum}}} and {{{ticket_custom}}} tables
   (appears to be a gratuitous change, but in the end, maybe the best solution...)

As neither solution satisfies me, I decided to postpone the renaming.

{{{category}}} was not that bad after all :)