TracCrossReferences branch: Wiki Formatter refactoring
The Wiki Formatter now optionally takes into account
the source object and the facet which corresponds to
the wiki text being formatted. So do the macros.

The issues #1802 and #1829 can therefore be properly fixed:
the source object is known and does not have to be guessed
from the path info.

As a bonus, macros like the PageOutline can now work for all Trac objects.