Experimenting with `ExtensionPoint` descriptors, and in particular introduced a `SingleExtensionPoint` which returns a single implementation for a given interface.
This was done by merging the proposal made by trac-form@swapoff.org in #2403
and what I already did in the FlexibleWikiPageNames branch, i.e.
warn if more than one implementation was active and suggest what
has to be changed in the TracIni file for only selecting one of them.

This has the advantage of re-using the normal way to enable/disable
components from within the TracIni.