The refactoring of TracCrossReferences continues...
The rebuilding of the cross-references is now done in the `xref.module`.

Consequently, a `ITracObjectManager` has now less to do:
It only has to provide the db type it manages (only one),
the factory method to create the subclass of `TracObject`
that it manages and an iterator over all existing objects
of that kind.

At the `TracObject` level, a bit more generic functionality
concerning the DB interface has been added:
 * a `_load` method that can be used to ensure a `TracObject`
   is in it's loaded form. `reload` should still be used if
   one wants to ''force'' reading from the DB.

The facets are now better handled at the level of each `TracObject`
(there will be soon a `WikiFacet` class again, I guess).