TracCrossReferences: improved cross-reference navigation.
Now, on the ''Cross-references'' page, when clicking on a reference
(an incoming or outgoing relation, or a backlink), one goes to the
''Cross-references'' page for that object, instead of targeting the
object itself. It's now the ''facet'' part of a reference that has
the link to the object's page (in the future, that link should even 
Target the facet's anchor within the object's page). 

That way, it's quite easy to browse through the graph of relations.